Select view for HTMLBars
